+++ src/main/java/kitchenpos/table/domain/OrderTableValidator.java
@@ -1,5 +1,5 @@
 package kitchenpos.table.domain;
 
-public interface OrderTableValidatable {
+public interface OrderTableValidator {
     void validateHasProgressOrder(OrderTable orderTable);
 }
